Fixed #698: Add "Claim" Button to VITA

+++ src/shared/wallet/account-section.tsx
@@ -5,29 +5,31 @@ import BigNumber from "bignumber.js";
 // @ts-ignore
 import window from "global/window";
 import { Account } from "iotex-antenna/lib/account/account";
-import { fromRau } from "iotex-antenna/lib/account/utils";
+import { fromRau, toRau } from "iotex-antenna/lib/account/utils";
 // @ts-ignore
 import { t } from "onefx/lib/iso-i18n";
 // @ts-ignore
 import { styled } from "onefx/lib/styletron-react";
 import React from "react";
 import { connect, DispatchProp } from "react-redux";
+import { RouteComponentProps, withRouter } from "react-router";
 import { AccountMeta } from "../../api-gateway/resolvers/antenna-types";
-import { ERC20Token, IERC20TokenInfoDict } from "../../erc20/erc20Token";
+import { ITokenInfo, ITokenInfoDict, Token } from "../../erc20/token";
 import { assetURL } from "../common/asset-url";
+import ConfirmContractModal from "../common/confirm-contract-modal";
 import { CopyButtonClipboardComponent } from "../common/copy-button-clipboard";
 import { onElectronClick } from "../common/on-electron-click";
 import { SpinPreloader } from "../common/spin-preloader";
 import { colors } from "../common/styles/style-color";
 import { TooltipButton } from "../common/tooltip-button";
 import { xconf, XConfKeys } from "../common/xconf";
-import AddCustomTokensFormModal from "./add-erc20-tokens-form-modal";
+import AddCustomTokensFormModal from "./add-custom-tokens-form-modal";
 import { ChainNetworkSwitch } from "./chain-network-switch";
 import { getAntenna } from "./get-antenna";
-import { setAccount, setERC20Tokens } from "./wallet-actions";
+import { setAccount, setTokens } from "./wallet-actions";
 import { IRPCProvider, IWalletState } from "./wallet-reducer";
 
-export interface Props extends DispatchProp {
+export interface Props extends DispatchProp, RouteComponentProps {
   account?: Account;
   createNew?: boolean;
   network?: IRPCProvider;
@@ -36,21 +38,27 @@ export interface Props extends DispatchProp {
 
 export interface State {
   accountMeta: AccountMeta | undefined;
-  erc20TokenInfos: IERC20TokenInfoDict;
+  tokenInfos: ITokenInfoDict;
   customTokensFormVisible: boolean;
   accountCheckID: string;
   isLoading: boolean;
+  isClaimingVita: boolean;
   isSyncing: boolean;
+  claimConfirmationVisible: boolean;
+  claimTokenAddress: string;
 }
 
 class AccountSection extends React.Component<Props, State> {
   public state: State = {
     accountMeta: undefined,
-    erc20TokenInfos: {},
+    tokenInfos: {},
     customTokensFormVisible: false,
     accountCheckID: "",
     isLoading: false,
-    isSyncing: false
+    isClaimingVita: false,
+    isSyncing: false,
+    claimConfirmationVisible: false,
+    claimTokenAddress: ""
   };
 
   private pollAccountInterval: number | undefined;
@@ -84,38 +92,36 @@ class AccountSection extends React.Component<Props, State> {
     const { account } = this.props;
     if (account) {
       await this.getAccount(account);
-      await this.getErc20TokensInfo();
+      await this.getTokensInfo();
       this.setState({ isLoading: false, isSyncing: false });
     }
     this.pollAccountInterval = window.setTimeout(this.pollAccount, 10000);
   };
 
-  public getErc20TokensInfo = async () => {
+  public getTokensInfo = async () => {
     const { account, dispatch, defaultNetworkTokens } = this.props;
     const { accountCheckID } = this.state;
     if (!account) {
       return;
     }
 
-    let erc20Addresses = xconf.getConf<Array<string>>(
-      `${XConfKeys.ERC20_TOKENS_ADDRS}_${accountCheckID}`,
+    let tokenAddresses = xconf.getConf<Array<string>>(
+      `${XConfKeys.TOKENS_ADDRS}_${accountCheckID}`,
       []
     );
 
-    erc20Addresses = [...defaultNetworkTokens, ...erc20Addresses];
+    tokenAddresses = [...defaultNetworkTokens, ...tokenAddresses];
     const tokenInfos = await Promise.all(
-      erc20Addresses.map(addr =>
-        ERC20Token.getToken(addr).getInfo(account.address)
-      )
+      tokenAddresses.map(addr => Token.getToken(addr).getInfo(account.address))
     );
-    const newErc20TokenInfos: IERC20TokenInfoDict = {};
+    const newTokenInfos: ITokenInfoDict = {};
     tokenInfos.forEach(info => {
       if (info && info.symbol) {
-        newErc20TokenInfos[info.erc20TokenAddress] = info;
+        newTokenInfos[info.tokenAddress] = info;
       }
     });
-    this.setState({ erc20TokenInfos: newErc20TokenInfos });
-    dispatch(setERC20Tokens(newErc20TokenInfos));
+    this.setState({ tokenInfos: newTokenInfos });
+    dispatch(setTokens(newTokenInfos));
   };
 
   public getAccount = async (account: Account) => {
@@ -172,10 +178,10 @@ class AccountSection extends React.Component<Props, State> {
     });
   };
 
-  public onAddCustomToken = async (erc20Address: string) => {
-    const { erc20TokenInfos } = this.state;
+  public onAddCustomToken = async (tokenAddress: string) => {
+    const { tokenInfos } = this.state;
     const { account, dispatch } = this.props;
-    if (erc20TokenInfos[erc20Address] || !account) {
+    if (tokenInfos[tokenAddress] || !account) {
       this.setState({
         customTokensFormVisible: false
       });
@@ -183,12 +189,12 @@ class AccountSection extends React.Component<Props, State> {
     }
 
     try {
-      const tokenInfo = await ERC20Token.getToken(erc20Address).getInfo(
+      const tokenInfo = await Token.getToken(tokenAddress).getInfo(
         account.address
       );
       if (!tokenInfo || !tokenInfo.symbol) {
         notification.error({
-          message: t("account.erc20.notfound"),
+          message: t("account.token.notfound"),
           duration: 3
         });
         this.setState({
@@ -196,41 +202,41 @@ class AccountSection extends React.Component<Props, State> {
         });
         return;
       }
-      erc20TokenInfos[erc20Address] = tokenInfo;
+      tokenInfos[tokenAddress] = tokenInfo;
     } catch (error) {
       notification.error({
-        message: t("account.erc20.addCustom"),
+        message: t("account.token.addCustom"),
         description: `${error.message}`,
         duration: 3
       });
     }
 
     xconf.setConf(
-      `${XConfKeys.ERC20_TOKENS_ADDRS}_${this.state.accountCheckID}`,
-      Object.keys(erc20TokenInfos)
+      `${XConfKeys.TOKENS_ADDRS}_${this.state.accountCheckID}`,
+      Object.keys(tokenInfos)
     );
     this.setState({
       customTokensFormVisible: false,
-      erc20TokenInfos: { ...erc20TokenInfos }
+      tokenInfos: { ...tokenInfos }
     });
-    dispatch(setERC20Tokens({ ...erc20TokenInfos }));
+    dispatch(setTokens({ ...tokenInfos }));
   };
 
-  public onDeleteErc20Token = (erc20Address: string) => {
+  public onDeleteErc20Token = (tokenAddress: string) => {
     const { dispatch, account } = this.props;
-    const { erc20TokenInfos } = this.state;
+    const { tokenInfos } = this.state;
 
-    if (!erc20TokenInfos[erc20Address]) {
+    if (!tokenInfos[tokenAddress]) {
       return;
     }
     // tslint:disable-next-line:no-dynamic-delete
-    delete erc20TokenInfos[erc20Address];
-    this.setState({ erc20TokenInfos: { ...erc20TokenInfos } });
-    dispatch(setERC20Tokens({ ...erc20TokenInfos }));
+    delete tokenInfos[tokenAddress];
+    this.setState({ tokenInfos: { ...tokenInfos } });
+    dispatch(setTokens({ ...tokenInfos }));
     if (account) {
       xconf.setConf(
-        `${XConfKeys.ERC20_TOKENS_ADDRS}_${this.state.accountCheckID}`,
-        Object.keys(erc20TokenInfos)
+        `${XConfKeys.TOKENS_ADDRS}_${this.state.accountCheckID}`,
+        Object.keys(tokenInfos)
       );
     }
   };
@@ -249,18 +255,85 @@ class AccountSection extends React.Component<Props, State> {
     );
   }
 
-  public renderBalance(): JSX.Element | null {
+  public claimVita = async (tokenAddress: string) => {
+    const { account, history } = this.props;
+    if (!account || !Token.getToken(tokenAddress).isVita()) {
+      return;
+    }
+    this.setState({
+      isClaimingVita: true
+    });
+    try {
+      const txHash = await Token.getToken(tokenAddress).claim(account);
+      // @ts-ignore
+      window.console.log(`Claimed VITA at action hash: ${txHash}`);
+      this.setState({
+        isClaimingVita: false
+      });
+      history.push(`/wallet/smart-contract/interact/${txHash}`);
+    } catch (e) {
+      notification.error({
+        message: `Failed to claim: ${e}`
+      });
+    }
+  };
+
+  private getColumns(): Array<Object> {
     const { defaultNetworkTokens } = this.props;
-    const { erc20TokenInfos, accountMeta, isLoading } = this.state;
-    const dataSource = Object.keys(erc20TokenInfos)
-      .map(addr => erc20TokenInfos[addr])
-      .filter(tokenInfo => tokenInfo && tokenInfo.symbol);
-    const columns = [
+    return [
       {
         title: "",
         dataIndex: "symbol",
         className: "wallet-token-symbol",
-        key: "symbol"
+        key: "symbol",
+        render: (
+          text: string,
+          token: ITokenInfo
+        ): JSX.Element | string | null => {
+          if (!Token.getToken(token.tokenAddress).isVita()) {
+            return text;
+          }
+          return (
+            <div style={{ paddingBottom: 42 }}>
+              <Row>{text}</Row>
+              <Row
+                type="flex"
+                justify="end"
+                align="middle"
+                style={{
+                  position: "absolute",
+                  padding: 10,
+                  width: "100%",
+                  margin: "0 -20px"
+                }}
+              >
+                <Button
+                  type="primary"
+                  href="https://discord.gg/4z3BTcd"
+                  style={{ marginRight: 10 }}
+                  target="_blank"
+                >
+                  {t("account.joinDiscord")}
+                </Button>
+                {
+                  // @ts-ignore
+                  <Button
+                    type="primary"
+                    loading={this.state.isClaimingVita}
+                    onClick={() => {
+                      this.setState({
+                        claimConfirmationVisible: true,
+                        claimTokenAddress: token.tokenAddress
+                      });
+                    }}
+                  >
+                    {t("account.claim")}
+                  </Button>
+                }
+              </Row>
+            </div>
+          );
+        }
       },
       {
         title: "",
@@ -270,11 +343,14 @@ class AccountSection extends React.Component<Props, State> {
       },
       {
         title: "",
-        dataIndex: "erc20TokenAddress",
+        dataIndex: "tokenAddress",
         className: "wallet-delete-token",
-        key: "erc20TokenAddress",
-        render: (text: string): JSX.Element | null =>
-          text && defaultNetworkTokens.indexOf(text) < 0 ? (
+        key: "tokenAddress",
+        render: (text: string): JSX.Element | null => {
+          if (!text || defaultNetworkTokens.indexOf(text) >= 0) {
+            return null;
+          }
+          return (
             <Icon
               type="close-circle"
               onClick={() => this.onDeleteErc20Token(text)}
@@ -283,14 +359,23 @@ class AccountSection extends React.Component<Props, State> {
                 cursor: "pointer"
               }}
             />
-          ) : null
+          );
+        }
       }
     ];
+  }
+
+  public renderBalance(): JSX.Element | null {
+    const { tokenInfos, accountMeta, isLoading } = this.state;
+    const dataSource = Object.keys(tokenInfos)
+      .map(addr => tokenInfos[addr])
+      .filter(tokenInfo => tokenInfo && tokenInfo.symbol);
+    const columns = this.getColumns();
     if (accountMeta) {
       dataSource.unshift({
         symbol: "IOTX",
         balanceString: fromRau(accountMeta.balance, ""),
-        erc20TokenAddress: "",
+        tokenAddress: "",
         balance: new BigNumber(accountMeta.balance),
         decimals: new BigNumber(0),
         name: "IOTX"
@@ -313,7 +398,6 @@ class AccountSection extends React.Component<Props, State> {
     if (!account) {
       return null;
     }
-
     return (
       <Card
         bodyStyle={{ padding: 0, wordBreak: "break-word" }}
@@ -343,7 +427,7 @@ class AccountSection extends React.Component<Props, State> {
               style={{ cursor: "pointer", fontSize: 13 }}
               onClick={() => this.showCustomTokensForm()}
             >
-              <Icon type="plus" /> {t("account.erc20.addCustom")}
+              <Icon type="plus" /> {t("account.token.addCustom")}
             </Col>
           </Row>
           <Row
@@ -405,7 +489,40 @@ class AccountSection extends React.Component<Props, State> {
           </Col>
         </Row>
         {this.renderCustomTokenForm()}
+        {this.renderClaimConfirmation()}
       </Card>
+    );
+  };
+
+  private readonly renderClaimConfirmation = (): JSX.Element | null => {
+    const { account } = this.props;
+    const { claimConfirmationVisible, claimTokenAddress } = this.state;
+    if (!account) {
+      return null;
+    }
+
+    const dataSource = {
+      address: account.address,
+      toContract: claimTokenAddress,
+      method: "claim",
+      amount: "0 IOTX",
+      limit: "100000",
+      price: toRau("1", "Qev")
+    };
+
+    return (
+      <ConfirmContractModal
+        dataSource={dataSource}
+        confirmContractOk={(ok: boolean) => {
+          if (ok) {
+            this.claimVita(claimTokenAddress);
+          }
+          this.setState({
+            claimConfirmationVisible: false
+          });
+        }}
+        showModal={claimConfirmationVisible}
+      />
     );
   };
 
@@ -435,4 +552,4 @@ const mapStateToProps = (state: {
   defaultNetworkTokens: (state.wallet || {}).defaultNetworkTokens || []
 });
 
-export default connect(mapStateToProps)(AccountSection);
+export default connect(mapStateToProps)(withRouter(AccountSection));
